benchmark: add cmake support
Additionally kept only one version of the benchmark program.
