+++ examples/benchmark/peanut_benchmark.c
@@ -1,67 +1,71 @@
+/**
+ * MIT License
+ * Copyright (c) 2018-2022 Mahyar Koshkouei
+ *
+ * Performs a benchmark of Peanut-GB with a specified ROM.
+ * Plays the ROM five times and prints the FPS for each play.
+ */
+#ifndef ENABLE_LCD
+# define ENABLE_LCD 1
+#endif
+
+/* Sound is disabled for this project. */
 #define ENABLE_SOUND 0
-#define ENABLE_LCD 0
 
 /* Import emulator library. */
 #include "../../peanut_gb.h"
 
 #include <errno.h>
-#include <stdint.h>
+#include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
 #include <time.h>
+
+const uint_fast32_t frames_per_run = 64 * 1024;
 
 struct priv_t
 {
 	/* Pointer to allocated memory holding GB file. */
 	uint8_t *rom;
+	/* Pointer to allocated memory holding save file. */
+	uint8_t *cart_ram;
+
+	/* Frame buffer */
+	uint16_t fb[LCD_HEIGHT][LCD_WIDTH];
 };
 
 /**
- * Return byte from blarrg test ROM.
+ * Returns a byte from the ROM file at the given address.
  */
-uint8_t gb_rom_read(struct gb_s *gb, const uint_fast32_t addr)
+static uint8_t gb_rom_read(struct gb_s *gb, const uint_fast32_t addr)
 {
 	const struct priv_t * const p = gb->direct.priv;
 	return p->rom[addr];
 }
 
 /**
- * Ignore cart RAM writes, since the test doesn't require it.
+ * Returns a byte from the cartridge RAM at the given address.
  */
-void gb_cart_ram_write(struct gb_s *gb, const uint_fast32_t addr, const uint8_t val)
+static uint8_t gb_cart_ram_read(struct gb_s *gb, const uint_fast32_t addr)
 {
-	(void)gb;
-	(void)addr;
-	(void)val;
-	return;
+	const struct priv_t * const p = gb->direct.priv;
+	return p->cart_ram[addr];
 }
 
 /**
- * Ignore cart RAM reads, since the test doesn't require it.
+ * Writes a given byte to the cartridge RAM at the given address.
  */
-uint8_t gb_cart_ram_read(struct gb_s *gb, const uint_fast32_t addr)
+static void gb_cart_ram_write(struct gb_s *gb, const uint_fast32_t addr,
+		const uint8_t val)
 {
-	(void)gb;
-	(void)addr;
-	return 0xFF;
-}
-
-/**
- * Ignore all errors.
- */
-void gb_error(struct gb_s *gb, const enum gb_error_e gb_err, const uint16_t val)
-{
-	(void)gb;
-	(void)gb_err;
-	(void)val;
-	return;
+	const struct priv_t * const p = gb->direct.priv;
+	p->cart_ram[addr] = val;
 }
 
 /**
  * Returns a pointer to the allocated space containing the ROM. Must be freed.
  */
-uint8_t *read_rom_to_ram(const char *file_name)
+static uint8_t *read_rom_to_ram(const char *file_name)
 {
 	FILE *rom_file = fopen(file_name, "rb");
 	size_t rom_size;
@@ -86,37 +90,76 @@ uint8_t *read_rom_to_ram(const char *file_name)
 	return rom;
 }
 
+/**
+ * Ignore all errors.
+ */
+static void gb_error(struct gb_s *gb, const enum gb_error_e gb_err, const uint16_t val)
+{
+	const char* gb_err_str[4] = {
+		"UNKNOWN",
+		"INVALID OPCODE",
+		"INVALID READ",
+		"INVALID WRITE"
+	};
+	fprintf(stderr, "Error %d occurred: %s\n. Abort.\n",
+			gb_err,
+			gb_err >= GB_INVALID_MAX ?
+				gb_err_str[0] : gb_err_str[gb_err]);
+
+	/* Unused parameters. */
+	(void)gb;
+	(void)val;
+
+	abort();
+}
+
+#if ENABLE_LCD
+/**
+ * Draws scanline into framebuffer.
+ */
+static void lcd_draw_line(struct gb_s *gb, const uint8_t pixels[160],
+		const uint_least8_t line)
+{
+	struct priv_t *priv = gb->direct.priv;
+	const uint16_t palette[] = { 0x7FFF, 0x5294, 0x294A, 0x0000 };
+
+	for (unsigned int x = 0; x < LCD_WIDTH; x++)
+		priv->fb[line][x] = palette[pixels[x] & 3];
+}
+#endif
+
 int main(int argc, char **argv)
 {
-	uint_fast32_t benchmark_ticks_total = 0;
-	uint_fast32_t benchmark_fps_total = 0;
-	struct priv_t priv = { .rom = NULL };
+	/* Must be freed */
+	char *rom_file_name = NULL;
 
-	if(argc != 2)
+	switch(argc)
 	{
-		printf("%s ROM\n", argv[0]);
-		exit(EXIT_FAILURE);
-	}
+		case 2:
+			rom_file_name = argv[1];
+			break;
 
-	/* Copy input ROM file to allocated memory. */
-	if((priv.rom = read_rom_to_ram(argv[1])) == NULL)
-	{
-		printf("Error(%d): %s\n", __LINE__, strerror(errno));
-		exit(EXIT_FAILURE);
+		default:
+			fprintf(stderr, "%s ROM\n", argv[0]);
+			exit(EXIT_FAILURE);
 	}
-
-	puts("Benchmark started");
 
 	for(unsigned int i = 0; i < 5; i++)
 	{
 		/* Start benchmark. */
 		struct gb_s gb;
-		const uint64_t start_time = (uint64_t)clock();
-		uint64_t bench_ticks;
-		uint64_t bench_fps;
-		uint64_t frames_to_run = 60 * 60 * 2; // 2 Minutes worth of frames
-		const uint64_t frames = frames_to_run;
-		int ret;
+		struct priv_t priv;
+
+		clock_t start_time;
+		uint_fast32_t frames = 0;
+		enum gb_init_error_e ret;
+
+		/* Copy input ROM file to allocated memory. */
+		if((priv.rom = read_rom_to_ram(rom_file_name)) == NULL)
+		{
+			printf("%d: %s\n", __LINE__, strerror(errno));
+			exit(EXIT_FAILURE);
+		}
 
 		/* Initialise context. */
 		ret = gb_init(&gb, &gb_rom_read, &gb_cart_ram_read,
@@ -124,28 +167,39 @@ int main(int argc, char **argv)
 
 		if(ret != GB_INIT_NO_ERROR)
 		{
-			printf("Error(%d): %d\n", __LINE__, ret);
+			fprintf(stderr, "Peanut-GB failed to initialise: %d\n",
+					ret);
 			exit(EXIT_FAILURE);
 		}
 
-		/* Step CPU until test is complete. */
+		printf("Run %u: ", i);
+		priv.cart_ram = malloc(gb_get_save_size(&gb));
+
+#if ENABLE_LCD
+		gb_init_lcd(&gb, &lcd_draw_line);
+		// gb.direct.interlace = 1;
+#endif
+
+		start_time = clock();
+
 		do
 		{
+			/* Execute CPU cycles until the screen has to be
+			 * redrawn. */
 			gb_run_frame(&gb);
-		} while(--frames_to_run);
+		}
+		while(++frames < frames_per_run);
 
-		/* End benchmark. */
-		bench_ticks = (uint64_t)clock() - start_time;
-		benchmark_ticks_total += bench_ticks;
+		{
+			double duration =
+				(double)(clock() - start_time) / CLOCKS_PER_SEC;
+			double fps = frames / duration;
+			printf("%f FPS, dur: %f\n", fps, duration);
+		}
 
-		bench_fps = frames / ((double)bench_ticks / CLOCKS_PER_SEC);
-		benchmark_fps_total += bench_fps;
-
-		printf("Benchmark %i: %llu\tFPS: %llu\n", i, bench_ticks, bench_fps);
+		free(priv.cart_ram);
+		free(priv.rom);
 	}
 
-	printf("Average    : %d\tFPS: %d\n",
-			benchmark_ticks_total / 5, benchmark_fps_total / 5);
-
-	return 0;
+	return EXIT_SUCCESS;
 }
